test verification and posting service
