+++ src/components/about/Values/index.tsx
@@ -11,8 +11,8 @@ const data: MissionItem[] = [
     {
         icon: <svg width="40" height="40" viewBox="0 0 40 40" fill="none" xmlns="http://www.w3.org/2000/svg">
             <rect width="40" height="40" rx="20" fill="white" fill-opacity="0.13" />
-            <path d="M18 14L24 20L18 26M10 14L16 20L10 26" stroke="white" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" />
-            <path d="M25 14L31 20L25 26M17 14L23 20L17 26" stroke="white" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" />
+            <path d="M18 14L24 20L18 26M10 14L16 20L10 26" stroke="white" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" />
+            <path d="M25 14L31 20L25 26M17 14L23 20L17 26" stroke="white" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" />
         </svg>,
         title: "Swift",
         description: "We pride ourselves on rapid response times and quick service delivery."
@@ -29,8 +29,8 @@ const data: MissionItem[] = [
         icon: <svg width="40" height="40" viewBox="0 0 40 40" fill="none" xmlns="http://www.w3.org/2000/svg">
             <rect width="40" height="40" rx="20" fill="white" fill-opacity="0.13" />
             <mask id="mask0_3537_8297" maskUnits="userSpaceOnUse" x="8" y="7" width="25" height="25">
-                <path d="M31.1875 18.9375V13.875C31.1875 13.5766 31.069 13.2905 30.858 13.0795C30.647 12.8685 30.3609 12.75 30.0625 12.75H20.5L17.6875 9.375H10.9375C10.6391 9.375 10.353 9.49353 10.142 9.7045C9.93103 9.91548 9.8125 10.2016 9.8125 10.5V28.5C9.8125 28.7984 9.93103 29.0845 10.142 29.2955C10.353 29.5065 10.6391 29.625 10.9375 29.625H19.375" stroke="white" stroke-width="3" stroke-linecap="round" stroke-linejoin="round" />
-                <path d="M23.7625 22.3125H28.4875L30.0625 24.2981L26.125 29.0625L22.1875 24.2981L23.7625 22.3125Z" fill="#555555" stroke="white" stroke-width="3" stroke-linecap="round" stroke-linejoin="round" />
+                <path d="M31.1875 18.9375V13.875C31.1875 13.5766 31.069 13.2905 30.858 13.0795C30.647 12.8685 30.3609 12.75 30.0625 12.75H20.5L17.6875 9.375H10.9375C10.6391 9.375 10.353 9.49353 10.142 9.7045C9.93103 9.91548 9.8125 10.2016 9.8125 10.5V28.5C9.8125 28.7984 9.93103 29.0845 10.142 29.2955C10.353 29.5065 10.6391 29.625 10.9375 29.625H19.375" stroke="white" strokeWidth="3" strokeLinecap="round" strokeLinejoin="round" />
+                <path d="M23.7625 22.3125H28.4875L30.0625 24.2981L26.125 29.0625L22.1875 24.2981L23.7625 22.3125Z" fill="#555555" stroke="white" strokeWidth="3" strokeLinecap="round" strokeLinejoin="round" />
             </mask>
             <g mask="url(#mask0_3537_8297)">
                 <path d="M7 6H34V33H7V6Z" fill="white" />
@@ -67,7 +67,7 @@ const data: MissionItem[] = [
     {
         icon: <svg width="40" height="40" viewBox="0 0 40 40" fill="none" xmlns="http://www.w3.org/2000/svg">
             <rect width="40" height="40" rx="20" fill="white" fill-opacity="0.13" />
-            <path d="M18 8C13.582 8 10 11.545 10 15.919C10 17.412 10.417 18.809 11.143 20M21.5 17V19M24.5 17V19M21.5 9V11M24.5 9V11M20 12.5H18M20 15.5H18M28 12.5H26M28 15.5H26M21 25L20.87 25.647C20.73 26.354 20.659 26.707 20.5 26.987C20.255 27.4186 19.8583 27.7436 19.387 27.899C19.082 28 18.72 28 18 28C17.28 28 16.918 28 16.613 27.9C16.1415 27.7444 15.7448 27.4189 15.5 26.987C15.341 26.707 15.27 26.354 15.13 25.647L15 25M24 11H22C21.057 11 20.586 11 20.293 11.293C20 11.586 20 12.057 20 13V15C20 15.943 20 16.414 20.293 16.707C20.586 17 21.057 17 22 17H24C24.943 17 25.414 17 25.707 16.707C26 16.414 26 15.943 26 15V13C26 12.057 26 11.586 25.707 11.293C25.414 11 24.943 11 24 11ZM13.383 23.098C13.291 22.822 13.245 22.683 13.25 22.571C13.2557 22.4552 13.2947 22.3435 13.3625 22.2494C13.4303 22.1554 13.5239 22.083 13.632 22.041C13.736 22 13.882 22 14.172 22H21.828C22.119 22 22.264 22 22.368 22.04C22.4762 22.0821 22.5699 22.1546 22.6377 22.2489C22.7055 22.3431 22.7445 22.455 22.75 22.571C22.755 22.683 22.709 22.821 22.617 23.098C22.447 23.609 22.362 23.865 22.231 24.072C21.957 24.5046 21.5275 24.8156 21.031 24.941C20.793 25 20.525 25 19.988 25H16.012C15.475 25 15.206 25 14.969 24.94C14.4727 24.8149 14.0432 24.5042 13.769 24.072C13.638 23.865 13.553 23.609 13.383 23.098Z" stroke="white" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" />
+            <path d="M18 8C13.582 8 10 11.545 10 15.919C10 17.412 10.417 18.809 11.143 20M21.5 17V19M24.5 17V19M21.5 9V11M24.5 9V11M20 12.5H18M20 15.5H18M28 12.5H26M28 15.5H26M21 25L20.87 25.647C20.73 26.354 20.659 26.707 20.5 26.987C20.255 27.4186 19.8583 27.7436 19.387 27.899C19.082 28 18.72 28 18 28C17.28 28 16.918 28 16.613 27.9C16.1415 27.7444 15.7448 27.4189 15.5 26.987C15.341 26.707 15.27 26.354 15.13 25.647L15 25M24 11H22C21.057 11 20.586 11 20.293 11.293C20 11.586 20 12.057 20 13V15C20 15.943 20 16.414 20.293 16.707C20.586 17 21.057 17 22 17H24C24.943 17 25.414 17 25.707 16.707C26 16.414 26 15.943 26 15V13C26 12.057 26 11.586 25.707 11.293C25.414 11 24.943 11 24 11ZM13.383 23.098C13.291 22.822 13.245 22.683 13.25 22.571C13.2557 22.4552 13.2947 22.3435 13.3625 22.2494C13.4303 22.1554 13.5239 22.083 13.632 22.041C13.736 22 13.882 22 14.172 22H21.828C22.119 22 22.264 22 22.368 22.04C22.4762 22.0821 22.5699 22.1546 22.6377 22.2489C22.7055 22.3431 22.7445 22.455 22.75 22.571C22.755 22.683 22.709 22.821 22.617 23.098C22.447 23.609 22.362 23.865 22.231 24.072C21.957 24.5046 21.5275 24.8156 21.031 24.941C20.793 25 20.525 25 19.988 25H16.012C15.475 25 15.206 25 14.969 24.94C14.4727 24.8149 14.0432 24.5042 13.769 24.072C13.638 23.865 13.553 23.609 13.383 23.098Z" stroke="white" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" />
         </svg>,
         title: "Innovation",
         description: "We continually innovate to improve user experiences and outcomes."
